перенес GetShopOwner в репозиторий user, обновил интерфейс репозитория в service, передал в зависимость offerService userRepository
